feat(backend): [PYTA-117] Query utils refactor
