+++ backend/app/core/utils/stock_utils.py
@@ -1,144 +1,24 @@
-import hashlib
-
 from collections.abc import Sequence
 from typing import Any
 
-import bcrypt
-
-from sqlalchemy import Row, RowMapping, func, update
+from sqlalchemy import Row, RowMapping, func
 from sqlmodel import Session, and_, or_, select
 
 from app.core.constants.time import UNIT_TIME
 from app.core.exceptions import (
-	EmailAlreadyExistsError,
-	InvalidCredentialsError,
 	TickerNotFoundError,
-	UserCreationError,
-	UserNotFoundError,
 )
-from app.core.models.auth import Auth
 from app.core.models.company import Company
 from app.core.models.exchange import Exchange
 from app.core.models.industry import Industry
-from app.core.models.portfolio import Portfolio
-from app.core.models.role import Role, RoleType
-from app.core.models.session import Session as SessionModel
 from app.core.models.stock_history import StockHistory
-from app.core.models.user import User
-from app.core.schemas.user_login import UserLogin
-from app.core.schemas.user_out import UserOut
-from app.core.schemas.user_register import UserRegister
 from app.core.utils.datetime_utils import DateTimeUtils
 from app.database import database_manager
 
 db = database_manager
 
 
-class QueryingUtils:
-	@staticmethod
-	def get_default_role_id(session: Session) -> str:
-		if default_role := session.exec(select(Role).where(Role.role == RoleType.USER)).first():
-			return default_role.id
-		else:  # pragma: no cover
-			raise ValueError("Default role not found")
-
-	@staticmethod
-	def register(user: UserRegister, ip: str) -> str:
-		with db.get_session() as session:
-			if session.exec(select(Auth).where(Auth.email == user.email)).first():
-				raise EmailAlreadyExistsError
-
-			new_auth = Auth(
-				email=user.email,
-				password=user.password,
-				role_id=QueryingUtils.get_default_role_id(session),
-			)
-			session.add(new_auth)
-
-			new_user = User(
-				name=user.name,
-				last_name=user.last_name,
-				date_of_birth=user.date_of_birth,
-				auth_id=new_auth.id,
-			)
-			session.add(new_user)
-			session.flush()
-
-			new_session: SessionModel = SessionModel(user_id=new_user.id, device_ip=ip)
-			session.add(new_session)
-
-			try:
-				session.commit()
-			except Exception as _:  # pragma: no cover
-				raise UserCreationError from None
-
-			return new_session.id
-
-	@staticmethod
-	def login(user_req: UserLogin, ip: str) -> str:
-		with db.get_session() as session:
-			auth = session.exec(select(Auth).where(Auth.email == user_req.email)).first()
-
-			if not auth or not bcrypt.checkpw(
-				user_req.password.encode("utf-8"), auth.password.encode("utf-8")
-			):
-				raise InvalidCredentialsError
-
-			user = session.exec(select(User).where(User.auth_id == auth.id)).first()
-
-			if not user:  # pragma: no cover
-				raise UserNotFoundError
-
-			new_session: SessionModel = SessionModel(user_id=user.id, device_ip=ip)
-
-			session.add(new_session)
-			session.commit()
-
-			return new_session.id
-
-	@staticmethod
-	def logout(session_id: str) -> None:
-		with db.get_session() as session:
-			if user_session := session.exec(
-				select(SessionModel).where(SessionModel.id == session_id)
-			).first():
-				session.delete(user_session)
-			else:
-				return
-
-	@staticmethod
-	def get_user_info(session_id: str) -> UserOut:
-		with db.get_session() as session:
-			session_data = session.exec(
-				select(SessionModel).where(SessionModel.id == session_id)
-			).first()
-			if not session_data:
-				raise UserNotFoundError
-
-			user = session.exec(select(User).where(User.id == session_data.user_id)).first()
-			if not user:
-				raise UserNotFoundError
-
-			if auth := session.exec(select(Auth).where(Auth.id == user.auth_id)).first():
-				return UserOut(
-					first_name=user.name,
-					last_name=user.last_name,
-					hashed_email=hashlib.sha256(auth.email.encode()).hexdigest(),
-					balance=user.balance,
-				)
-			else:
-				raise UserNotFoundError
-
-	@staticmethod
-	def get_user_from_token(session: Session, session_id: str) -> dict[str, Any] | None:
-		user = session.exec(
-			select(User)
-			.where(SessionModel.id == session_id)
-			.join(SessionModel, SessionModel.user_id == User.id)  # type: ignore[arg-type]
-		).first()
-
-		return dict(user) if user else None
-
+class StockUtils:
 	@staticmethod
 	def get_stocks() -> list[dict[Any, Any]]:
 		with db.get_session() as session:
@@ -322,61 +202,3 @@ class QueryingUtils:
 			)
 		)
 		return session.execute(query).mappings().all()
-
-	@staticmethod
-	def get_user_balance(session: Session) -> float:
-		if balance := session.exec(select(User.balance).first()):  # type: ignore[attr-defined]
-			return balance  # type: ignore[return-value]
-		else:
-			raise UserNotFoundError
-
-	@staticmethod
-	def update_user_balance(session: Session, user_id: str, change: float) -> None:
-		session.execute(
-			update(User).where(User.id == user_id).values(balance=User.balance + change)  # type: ignore[arg-type]
-		)
-
-	@staticmethod
-	def get_or_create_portfolio(session: Session, user_id: str, company_id: str) -> Portfolio:
-		if portfolio := session.exec(
-			select(Portfolio).where(
-				and_(Portfolio.user_id == user_id, Portfolio.company_id == company_id)
-			)
-		).first():
-			return portfolio
-
-		new_portfolio = Portfolio(user_id=user_id, company_id=company_id, amount=0)
-		session.add(new_portfolio)
-		return new_portfolio
-
-	@staticmethod
-	def update_user_portfolio(
-		session: Session, user_id: str, company_id: str, amount: float, buy: bool
-	) -> None:
-		portfolio = QueryingUtils.get_or_create_portfolio(session, user_id, company_id)
-
-		portfolio.amount += amount if buy else -amount
-
-		session.add(portfolio)
-
-	@staticmethod
-	def get_user_portfolio(session: Session, user_id: str, stock_id: str) -> Portfolio | None:
-		query = select(Portfolio).where(
-			and_(Portfolio.user_id == user_id, Portfolio.company_id == stock_id)
-		)
-
-		return session.exec(query).one_or_none()
-
-	@staticmethod
-	def get_user_portfolios(session: Session, user_id: str) -> Sequence[Row[Any]]:
-		query = (
-			select(
-				Portfolio,
-				Company,
-			)
-		).where(
-			Portfolio.user_id == user_id,
-			Portfolio.company_id == Company.id,
-			Portfolio.amount > 0,
-		)
-		return session.execute(query).all()
